support for skipping Pandora stations
Support added for skipping stations in Pandora. Skips to the next
station or to the top of the list if it reached the bottom. Skips over
the shuffle button, because when shuffle was pressed it changed the
display of the stations and prevented further skipping. I used the
fast-forward icon, but need to come up with a better icon.